# packages/contracts/m1-identity.openapi.yaml
openapi: "3.1.0"
info:
  title: M1 Identity
  version: "1.0.0"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenResponse:
      type: object
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        token_type: {type: string, example: bearer}
    UserOut:
      type: object
      properties:
        id: {type: integer}
        email: {type: string}
        name: {type: string}
        is_active: {type: boolean}
        role_id: {type: integer, nullable: true}
        department_id: {type: integer, nullable: true}
    SyncReport:
      type: object
      properties:
        roles_added: {type: integer}
        roles_updated: {type: integer}
        departments_added: {type: integer}
        departments_updated: {type: integer}
        users_added: {type: integer}
        users_updated: {type: integer}
        users_deactivated: {type: integer}
        errors: {type: array, items: {type: string}}
        duration_ms: {type: number}
    SyncRunOut:
      type: object
      properties:
        id: {type: integer}
        started_at: {type: string, format: date-time, nullable: true}
        finished_at: {type: string, format: date-time, nullable: true}
        status: {type: string, enum: [queued, running, success, error]}
        report:
          $ref: "#/components/schemas/SyncReport"
          nullable: true
        error: {type: string, nullable: true}
    SyncHistoryResponse:
      type: array
      items:
        $ref: "#/components/schemas/SyncRunOut"
paths:
...
  /auth/login:
    post:
      summary: Email+password login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string, format: email}
                password: {type: string}
      responses:
        "200":
          description: JWT tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401": {description: Invalid credentials}
        "429": {description: Locked out (too many failures), headers: {Retry-After: {schema: {type: integer}}}}
  /auth/refresh:
    post:
      summary: Rotate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401": {description: Invalid / replayed token}
  /auth/logout:
    post:
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        "204": {description: Logged out}
  /auth/me:
    get:
      summary: Current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile with role + permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: {type: integer}
                  email: {type: string}
                  name: {type: string}
                  role: {type: string}
                  permissions: {type: array, items: {type: string}}
        "401": {description: Unauthorized}
  /users:
    get:
      summary: List users (admin only, paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: size
          in: query
          schema: {type: integer, default: 50, maximum: 200}
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserOut"
        "403": {description: Forbidden}
    post:
      summary: Create user (admin.write)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: {type: string}
                password: {type: string}
                name: {type: string}
                role_id: {type: integer}
                department_id: {type: integer}
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
  /users/{id}:
    get:
      summary: Get user by ID (user.read or self)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
        "403": {description: Forbidden}
        "404": {description: Not found}
    patch:
      summary: Update user (user.write or self for limited fields)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string}
                department_id: {type: integer}
                role_id: {type: integer}
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
    delete:
      summary: Soft-delete user (user.delete)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
      responses:
        "204": {description: Deleted (is_active=false)}
        "403": {description: Forbidden}
  /admin/sync/mysql:
    post:
      summary: Trigger MySQL sync (admin.write)
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Sync accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string, example: accepted}
                  run_id: {type: integer}
        "403": {description: Forbidden}
  /admin/sync/mysql/status:
    get:
      summary: Last sync run + next schedule (admin.read)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sync status
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_run:
                    $ref: "#/components/schemas/SyncRunOut"
                    nullable: true
                  sync_interval_minutes: {type: integer}
        "403": {description: Forbidden}
  /admin/sync/mysql/history:
    get:
      summary: Recent sync history (audit.read)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: {type: integer, default: 50, maximum: 200}
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncHistoryResponse"
        "403": {description: Forbidden}